refactor: rename and reword sync rewrite docstrings
Merge the Phase-2 dispatch test file into tests/services/test_saves.py
per the test-mirrors-source convention. Rename internal class names to
neutral form (e.g. internal Test classes for the new compute-action
path renamed to TestGetSaveStatusComputeAction).

Rewrite docstrings in sync_action.py module, saves.py service methods
(_sync_rom_saves, _get_save_status_io, resolve_sync_conflict), and
test module headers — describe what the code does and why, not what
inspired it.

Reword skip-mark reasons in 19 superseded tests across test_saves.py
and test_plugin_saves.py to reference the rewrite by what it is, not
by inspiration source.

1765 passed, 19 skipped. No production logic changes.
